pyproject.toml
[project]
name = "tdsp2-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncio==3.4.3",
    "beautifulsoup4==4.13.3",
    "certifi==2025.1.31",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi==0.115.11",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel>=6.29.5",
    "jellyfish==1.1.3",
    "jiter==0.9.0",
    "numpy==2.2.4",
    "openai==1.67.0",
    "openpyxl==3.1.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pypdf2==3.0.1",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "pytz==2025.1",
    "requests==2.32.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "starlette==0.46.1",
    "tabula-py==2.10.0",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "zipfile36==0.1.3",
]
